Azelia/requirements.txt
argon2-cffi==21.1.0
bcrypt==3.2.0
mysql-connector-python==8.0.27
re==2.2.1